feat: refactored into plugin folder and migrated from lsp-zero and installed tabout plugin
