--- after/plugin/autopairs.rc.lua
@@ -1,6 +1,0 @@
-local status, autopairs = pcall(require, 'nvim-autopairs')
-if (not status) then return end
-
-autopairs.setup {
-    disable_filetype = { 'TelescopePrompt', 'vim' }
-}
