--- after/plugin/colors.lua
@@ -1,11 +1,0 @@
-function setColor(color)
-    color = color or "everforest"
-	vim.cmd.colorscheme(color)
-
-	vim.api.nvim_set_hl(0, "Normal", { bg = "none" })	
-	--vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none" })
-end
-
-setColor()
-
-
