--- after/plugin/lazygit.lua
@@ -1,1 +1,0 @@
-vim.keymap.set("n", "<leader>gt", vim.cmd.LazyGit)
