--- after/plugin/ts-autotag.rc.lua
@@ -1,4 +1,0 @@
-local status, autotag = pcall(require, 'nvim-ts-autotag')
-if (not status) then return end
-
-autotag.setup {}
