+++ lua/0xMiyu/plugins/autopairs.lua
@@ -1,0 +1,9 @@
+return {
+	"windwp/nvim-autopairs",
+	event = "InsertEnter",
+	config = function()
+		local npairs = require("nvim-autopairs")
+		npairs.setup({})
+	end,
+	opts = {},
+}
