+++ lua/0xMiyu/plugins/barbar.lua
@@ -1,0 +1,51 @@
+-- barbar tabs
+return {
+	{
+		"romgrk/barbar.nvim",
+		dependencies = {
+			"lewis6991/gitsigns.nvim", -- OPTIONAL: for git status
+			"nvim-tree/nvim-web-devicons", -- OPTIONAL: for file icons
+		},
+		init = function()
+			vim.g.barbar_auto_setup = false
+		end,
+		opts = {
+			animation = false,
+			clickable = true,
+			icons = {
+				buffer_index = false,
+				buffer_number = false,
+				button = "",
+				diagnostics = {
+					[vim.diagnostic.severity.ERROR] = { enabled = true, icon = "ﬀ" },
+					[vim.diagnostic.severity.WARN] = { enabled = false },
+					[vim.diagnostic.severity.INFO] = { enabled = false },
+					[vim.diagnostic.severity.HINT] = { enabled = true },
+				},
+				gitsigns = {
+					added = { enabled = true, icon = "+" },
+					changed = { enabled = true, icon = "~" },
+					deleted = { enabled = true, icon = "-" },
+				},
+				filetype = {
+					custom_colors = true,
+					enabled = true,
+				},
+				separator = { left = "▎", right = "" },
+				modified = { button = "●" },
+				pinned = { button = "", filename = true },
+				preset = "default",
+				alternate = { filetype = { enabled = false } },
+				current = { buffer_index = true },
+				inactive = { button = "×" },
+				visible = { modified = { buffer_number = false } },
+			},
+			sidebar_filetypes = {
+				NvimTree = true,
+				undotree = { text = "undotree" },
+				["neo-tree"] = { event = "BufWipeout" },
+				Outline = { event = "BufWinLeave", text = "symbols-outline" },
+			},
+		},
+	},
+}
