+++ lua/0xMiyu/plugins/lazygit.lua
@@ -1,0 +1,13 @@
+-- lazygit
+return {
+	{
+		"kdheepak/lazygit.nvim",
+		-- optional for floating window border decoration
+		dependencies = {
+			"nvim-lua/plenary.nvim",
+		},
+		config = function()
+			vim.keymap.set("n", "<leader>gt", vim.cmd.LazyGit)
+		end,
+	},
+}
