+++ lua/0xMiyu/plugins/telescope.lua
@@ -1,0 +1,26 @@
+-- telescope.nvim
+return {
+	{
+		"nvim-telescope/telescope.nvim",
+		tag = "0.1.8",
+		-- or                            , branch = '0.1.x',
+		dependencies = { { "nvim-lua/plenary.nvim" } },
+		config = function()
+			require("telescope").setup({
+				pickers = {
+					find_files = {
+						-- `hidden = true` will still show the inside of `.git/` as it's not `.gitignore`d.
+						find_command = { "rg", "--files", "--hidden", "--glob", "!**/.git/*" },
+					},
+				},
+			})
+
+			local builtin = require("telescope.builtin")
+			vim.keymap.set("n", "<leader>ff", builtin.find_files, {})
+			vim.keymap.set("n", "<leader>fg", builtin.git_files, {})
+			vim.keymap.set("n", "<leader>fs", function()
+				builtin.grep_string({ search = vim.fn.input("Grep > ") })
+			end)
+		end,
+	},
+}
