+++ lua/0xMiyu/plugins/treesitter.lua
@@ -1,0 +1,30 @@
+-- treesitter syntax highlighting
+return {
+	{
+		"nvim-treesitter/nvim-treesitter",
+		build = ":TSUpdate",
+		opts = {
+			-- A list of parser names, or "all" (the five listed parsers should always be installed)
+			ensure_installed = {
+				"c",
+				"lua",
+				"vim",
+				"vimdoc",
+				"query",
+				"rust",
+				"javascript",
+				"typescript",
+				"java",
+				"python",
+			},
+
+			highlight = {
+				enable = true,
+				additional_vim_regex_highlighting = false,
+			},
+		},
+		config = function(_, opts)
+			require("nvim-treesitter.configs").setup(opts)
+		end,
+	},
+}
